+++ example/windows/flutter/generated_plugins.cmake
@@ -7,7 +7,8 @@ list(APPEND FLUTTER_PLUGIN_LIST
 )
 
 list(APPEND FLUTTER_FFI_PLUGIN_LIST
-  bergamot_translator
+  jni
+  mariannmt_wrapper
 )
 
 set(PLUGIN_BUNDLED_LIBRARIES)